Changed queue assign logic to queue rsj_req by prev_seq rather than by seq.
The practical result of this is that the ordering of rsj_reqs on a channel
is preserved.  There are arguments for both modes of behaviour (order
preserving and reordering), but since the ordering mode is less surprising
I've settled on that for now.  The possibility exists in the future to
permit a flag in the oob_data to select reordering behaviour for the use
cases that need it.